added services for crud operations and their interfaces

+++ MedicalRecordProject/src/main/java/org/example/medicalrecordproject/repositories/DiagnosisRepository.java
@@ -3,5 +3,5 @@ package org.example.medicalrecordproject.repositories;
 import org.example.medicalrecordproject.models.Diagnosis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiagnosisRepository extends JpaRepository<Diagnosis, Integer> {
+public interface DiagnosisRepository extends JpaRepository<Diagnosis, Long> {
 }
